requirements.txt
google-cloud-bigquery>=3.25.0
google-auth>=2.29.0
pyarrow>=15.0.0
pyyaml>=6.0.1
pandas>=2.2.0
